+++ src/Surfnet/Stepup/Configuration/Event/SelectRaaOptionChangedEvent.php
@@ -1,0 +1,70 @@
+<?php
+
+/**
+ * Copyright 2018 SURFnet B.V.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+namespace Surfnet\Stepup\Configuration\Event;
+
+use Broadway\Serializer\SerializableInterface;
+use Surfnet\Stepup\Configuration\Value\Institution;
+use Surfnet\Stepup\Configuration\Value\InstitutionConfigurationId;
+use Surfnet\Stepup\Configuration\Value\SelectRaaOption;
+
+final class SelectRaaOptionChangedEvent implements SerializableInterface
+{
+    /**
+     * @var InstitutionConfigurationId
+     */
+    public $institutionConfigurationId;
+
+    /**
+     * @var Institution
+     */
+    public $institution;
+
+    /**
+     * @var SelectRaaOption
+     */
+    public $selectRaaOption;
+
+    public function __construct(
+        InstitutionConfigurationId $institutionConfigurationId,
+        Institution $institution,
+        SelectRaaOption $selectRaaOption
+    ) {
+        $this->institutionConfigurationId = $institutionConfigurationId;
+        $this->institution = $institution;
+        $this->selectRaaOption = $selectRaaOption;
+    }
+
+    public static function deserialize(array $data)
+    {
+        return new self(
+            new InstitutionConfigurationId($data['institution_configuration_id']),
+            new Institution($data['institution']),
+            new SelectRaaOption($data['select_raa_option'])
+        );
+    }
+
+    public function serialize()
+    {
+        return [
+            'institution_configuration_id' => $this->institutionConfigurationId->getInstitutionConfigurationId(),
+            'institution' => $this->institution->getInstitution(),
+            'select_raa_option' => $this->selectRaaOption->getInstitutions(),
+        ];
+    }
+}