+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # Changelog
 
+## Develop
+**New features**
+* New institution configuration options can be configured (useRa, useRaa and selectRaa) #232 #233
+* The previously hardcoded "server_version" config option (Doctrine DBAL) is now configurable
+
+## 2.9.1
+**Bugfix**
+* Resolve a Doctrine DBAL configuration issue described in https://github.com/doctrine/DoctrineBundle/issues/351
+
 ## 2.9.0
 **Bugfix**
 * Fix SRAA command for institution without config #228